Trying to deal with weird Taxonomy data
Files aren't currently working. There's an issue with the
gbif_species_in_genus() function.

Will try again in the morning to get it working but wanted a record of
today's updates.

Need to get this working so that we have a succinct list of only fish
species. Currently the location validation is taking too long / too
large of a file because it's pulling in bird data (which have
exponentially more observations than fish).