Improved the bayesian function: eDNA_OpenBUGS
Now uses beta priors instead of normally distributed priors. Posteriors are visibly more reasonable. There may be an issue running this for each row of the dataset, even if it only takes 1 second. Efficiency may be improved by moving the library(R2OpenBUGS), library(coda) and writeLines("") for OpenBUGS code to outside the function so they are only called once. Still, it may take a long time to run the simulations for each row