Merged all existing code into a single Rmd
This merges the processing that was occuring across several files into a
single Rmd file for ease of use.

This merging is a preparation step before the final merge, bringing in
Sweta's OBIS location data processing code.

Also core to this update is minimising the exporting of CSV files to the
data folder. All data is now held within R for processing and isn't
written back out until the very end to the file "data_enriched.csv".